Moved common hooks into separate file.
